Split dumper and scanner features
