+++ dumper/src/main.rs
@@ -1,8 +1,10 @@
 use dumper::{CommandEnum, Commands};
 
-fn main() -> Result<(), dumper::Error> {
-    match argh::from_env::<Commands>().cmds {
+fn main() {
+    if let Err(err) = match argh::from_env::<Commands>().cmds {
         CommandEnum::WithDisk(this) => this.init(),
         CommandEnum::TestPtrs(this) => this.init(),
+    } {
+        eprintln!("{err}")
     }
 }
