+++ ffi/Cargo.toml
@@ -10,6 +10,7 @@ crate-type = ["cdylib"]
 [dependencies.ptrsx]
 path = "../ptrsx"
 default-features = false
+features = ["dumper", "scanner"]
 
 [dependencies.vmmap]
 path = "../vmmap"
@@ -18,4 +19,3 @@ default-features = false
 [build-dependencies.cbindgen]
 version = "0.26.0"
 default-features = false
-
