+++ ptrsx/src/ptrsx_d.rs
@@ -1,8 +1,8 @@
-use std::{cmp::Ordering, collections::BTreeMap, io};
+use std::{cmp::Ordering, collections::BTreeMap, io, path::Path};
 
-use vmmap::VirtualMemoryRead;
+use vmmap::{Pid, Process, ProcessInfo, VirtualMemoryRead, VirtualQuery};
 
-use super::*;
+use super::{check_region, encode_modules, Error, Module, PtrsxScanner, DEFAULT_BUF_SIZE, PTRSIZE};
 
 pub fn create_pointer_map<P>(
     proc: &P,
@@ -125,4 +125,61 @@ where
     }
 
     Ok(())
+}
+
+impl PtrsxScanner {
+    pub fn create_pointer_map<P>(&mut self, proc: &P, is_align: bool) -> Result<(), Error>
+    where
+        P: ProcessInfo + VirtualMemoryRead,
+    {
+        let pages = proc.get_maps().filter(check_region).collect::<Vec<_>>();
+        let region = pages.iter().map(|m| (m.start(), m.size())).collect::<Vec<_>>();
+        let mut iter = pages.iter().flat_map(|m| {
+            let path = Path::new(m.name()?);
+            let name = path.has_root().then_some(path)?.to_str()?.to_string();
+            Some(Module { start: m.start(), end: m.end(), name })
+        });
+        let mut current = iter.next().ok_or("no base address module available.")?;
+        for page in iter {
+            if page.name == current.name {
+                current.end = page.end;
+            } else {
+                self.modules.push(current);
+                current = page;
+            }
+        }
+        self.modules.push(current);
+
+        self.forward = create_pointer_map(proc, &region, is_align)?;
+        for (&k, &v) in &self.forward {
+            self.reverse.entry(v).or_default().push(k);
+        }
+
+        Ok(())
+    }
+
+    pub fn create_pointer_map_file<W: io::Write>(&self, writer: &mut W, pid: Pid, is_align: bool) -> Result<(), Error> {
+        let proc = Process::open(pid)?;
+        let pages = proc.get_maps().filter(check_region).collect::<Vec<_>>();
+        let region = pages.iter().map(|m| (m.start(), m.size())).collect::<Vec<_>>();
+
+        let mut modules = Vec::with_capacity(pages.len());
+        let mut iter = pages.iter().flat_map(|m| {
+            let path = Path::new(m.name()?);
+            let name = path.has_root().then_some(path)?.to_str()?.to_string();
+            Some(Module { start: m.start(), end: m.end(), name })
+        });
+        let mut current = iter.next().ok_or("no base address module available.")?;
+        for page in iter {
+            if page.name == current.name {
+                current.end = page.end;
+            } else {
+                modules.push(current);
+                current = page;
+            }
+        }
+        modules.push(current);
+        encode_modules(&modules, writer)?;
+        create_pointer_map_with_writer(&proc, &region, is_align, writer)
+    }
 }
