+++ scanner/src/error.rs
@@ -1,4 +1,5 @@
-#[derive(Debug)]
+use std::fmt::Display;
+
 pub struct Error(pub String);
 
 impl From<ptrsx::Error> for Error {
@@ -28,5 +29,11 @@ impl From<String> for Error {
 impl From<std::num::ParseIntError> for Error {
     fn from(value: std::num::ParseIntError) -> Self {
         Self(value.to_string())
+    }
+}
+
+impl Display for Error {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "error: {}", self.0)
     }
 }
